Adds all

+++ FunGuy/Assets/Scripts/Game/Data/FungiData.cs
@@ -3,7 +3,7 @@ using Util;
 namespace Game.Data
 {
     public record FungiData(FungiState State = FungiState.Walking, bool IsOnGround = false, bool IsOnRoot = false,
-        int Health = GlobalConst.MaxHealth, int LastDamageTimeStamp = 0)
+        int Health = GlobalConst.MaxHealth, float LastDamageTimeStamp = 0f)
     {
         public bool CanJump => (IsOnGround || IsOnRoot) && State == FungiState.Walking;
